refactor(auth): split session model from platform contracts
