+++ ui/src/commonMain/kotlin/dev/obiente/nextcloudnative/app/NextcloudPlatform.kt
@@ -46,19 +46,6 @@ data class PlatformCapabilityStatus(
     val state: PlatformCapabilityState,
 )
 
-data class NextcloudSession(
-    val serverUrl: String,
-    val loginName: String,
-    val appPassword: String,
-) {
-    /** Opaque, credential-free identity for account-scoped process state. */
-    val accountId: NextcloudAccountId
-        get() = deriveNextcloudAccountId(serverUrl, loginName)
-
-    override fun toString(): String =
-        "NextcloudSession(serverUrl=<redacted>, loginName=<redacted>, appPassword=<redacted>)"
-}
-
 data class LoginChallenge(
     val enteredServerUrl: String,
     val pollEndpoint: String,
